Move update user hooks to auth
