+++ client/packages/common/src/authentication/api/api.ts
@@ -132,5 +132,14 @@ export const getAuthQueries = (sdk: Sdk, t: TypedTFunction<LocaleKey>) => ({
         return { nodes: [] };
       }
     },
+    lastSuccessfulUserSync: async () => {
+      return (await sdk.lastSuccessfulUserSync()).lastSuccessfulUserSync
+        .lastSuccessfulSync;
+    },
+    updateUser: async () => {
+      const result = await sdk.updateUser();
+
+      return result.updateUser;
+    },
   },
 });
