+++ client/packages/common/src/authentication/api/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useRefreshToken';
 export * from './useLogin';
 export * from './useGetUserPermissions';
 export * from './useIsCentralServer';
+export * from './useUpdateUser';
