+++ client/packages/common/src/authentication/api/hooks/useAuthApi.ts
@@ -12,6 +12,7 @@ export const useAuthApi = () => {
     me: (token: string) => ['me', token] as const,
     isCentralServer: ['isCentralServer'] as const,
     refresh: (token: string) => ['refresh', token] as const,
+    userSync: () => ['userSync'] as const,
   };
 
   return { ...queries, keys };
