+++ client/packages/common/src/authentication/api/hooks/useUpdateUser.ts
@@ -1,0 +1,14 @@
+import {
+  useAuthApi,
+  useMutation,
+  useQueryClient,
+} from '@openmsupply-client/common';
+
+export const useUpdateUser = (token: string) => {
+  const api = useAuthApi();
+  const queryClient = useQueryClient();
+
+  return useMutation(api.get.updateUser, {
+    onSettled: () => queryClient.invalidateQueries(api.keys.me(token)),
+  });
+};
