+++ client/packages/common/src/authentication/api/operations.graphql
@@ -135,3 +135,37 @@ query storePreferences($storeId: String!) {
     id
   }
 }
+
+fragment UpdateUser on UpdateUserNode {
+  lastSuccessfulSync
+}
+
+mutation updateUser {
+  updateUser {
+    __typename
+    ... on UpdateUserNode {
+      ...UpdateUser
+    }
+    ... on UpdateUserError {
+      __typename
+      error {
+        ... on InvalidCredentials {
+          __typename
+          description
+        }
+        ... on ConnectionError {
+          __typename
+          description
+        }
+        description
+      }
+    }
+  }
+}
+
+query lastSuccessfulUserSync {
+  lastSuccessfulUserSync {
+    __typename
+    ...UpdateUser
+  }
+}
