+++ client/packages/system/src/Sync/api/hooks/index.ts
@@ -10,12 +10,10 @@ export const useSync = {
     initialise: Sync.useInitialiseSite,
     manualSync: Sync.useManualSync,
     update: Sync.useUpdateSyncSettings,
-    updateUser: Sync.useUpdateUser,
   },
   utils: {
     syncStatus: Utils.useSyncStatus,
     mutateSyncStatus: Utils.useMutateSyncStatus,
     syncInfo: Utils.useSyncInfo,
-    lastSuccessfulUserSync: Utils.useLastSuccessfulUserSync,
   },
 };
