+++ client/packages/system/src/Sync/api/hooks/sync/index.ts
@@ -1,11 +1,9 @@
 import { useUpdateSyncSettings } from './updateSyncSettings';
 import { useInitialiseSite } from './useInitialiseSite';
 import { useManualSync } from './useManualSync';
-import { useUpdateUser } from './useUpdateUser';
 
 export const Sync = {
   useInitialiseSite,
   useManualSync,
   useUpdateSyncSettings,
-  useUpdateUser,
 };
