+++ client/packages/system/src/Sync/api/hooks/utils/index.ts
@@ -1,12 +1,10 @@
 import { useSyncApi } from './useSyncApi';
 import { useSyncStatus, useMutateSyncStatus } from './useSyncStatus';
 import { useSyncInfo } from './useSyncInfo';
-import { useLastSuccessfulUserSync } from './useLastSuccessfulUserSync';
 
 export const Utils = {
   useSyncApi,
   useSyncStatus,
   useSyncInfo,
   useMutateSyncStatus,
-  useLastSuccessfulUserSync,
 };
